requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
aioredis==1.3.1
astor==0.8.0
astunparse==1.6.3
async-timeout==3.0.1
atari-py==0.2.6
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.9.1
blessings==1.7
cachetools==4.1.1
certifi==2020.06.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorama==0.4.3
colorful==0.5.4
configparser==5.0.1
contextvars==2.4
cycler==0.10.0
Cython==0.29.21
deepdiff==4.3.2
dill==0.3.2
docker-pycreds==0.4.0
docopt==0.6.2
fasteners==0.15
filelock==3.0.12
funcsigs==1.0.2
future==0.16.0
gast==0.2.2
gin==0.1.6
gin-config==0.3.0
gitdb==4.0.5
GitPython==3.1.9
glfw==1.12.0
google==3.0.0
google-api-core==1.22.1
google-auth==1.21.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
gpustat==0.6.0
gql==0.2.0
graphql-core==1.1
grpcio==1.31.0
gym==0.17.2
h5py==2.10.0
hiredis==1.1.0
idna==2.7
idna-ssl==1.1.0
imageio==2.4.1
immutables==0.14
importlib-metadata==4.10.0
joblib==0.16.0
jsonnet==0.16.0
jsonpickle==0.9.6
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.0.1
lockfile==0.12.2
Markdown==3.1.1
matplotlib==3.3.1
mock==2.0.0
monotonic==1.5
more-itertools==4.3.0
mpyq==0.2.5
msgpack==1.0.0
multidict==4.7.6
munch==2.3.2
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencensus==0.7.10
opencensus-context==0.1.1
opencv-python==4.2.0.34
opt-einsum==3.1.0
ordered-set==4.0.2
packaging>=21.0,<25.0
pandas==1.1.1
pathlib2==2.3.2
pathtools==0.1.2
pbr==4.3.0
Pillow==6.2.1
pluggy==0.7.1
portpicker==1.2.0
probscale==0.2.3
progressbar2==3.53.1
prometheus-client==0.8.0
promise==2.3
protobuf==3.12.4
psutil==5.7.2
py==1.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyglet==1.5.0
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
pyparsing==2.2.2
pyrsistent==0.16.0
PySC2==3.0.0
pytest==3.8.2
python-dateutil==2.7.3
python-utils==2.4.0
pytz==2020.1
PyYAML==3.13
pyzmq==19.0.2
redis==3.4.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s2clientprotocol==4.10.1.75800.0
s2protocol==4.11.4.78285.0
sacred==0.7.2
scipy==1.4.1
seaborn==0.10.1
sentry-sdk==0.18.0
setproctitle==1.1.10
shortuuid==1.0.1
six==1.15.0
sk-video==1.1.10
smmap==3.0.4
snakeviz==1.0.0
soupsieve==2.0.1
subprocess32==3.5.4
tabulate==0.8.7
termcolor==1.1.0
tensorboardX
tornado==5.1.1
tqdm==4.48.2
typing-extensions==3.7.4.3
urllib3==1.23
wandb==0.10.5
watchdog==0.10.3
websocket-client==0.53.0
Werkzeug==0.16.1
whichcraft==0.5.2
wrapt==1.12.1
xmltodict==0.12.0
yarl==1.5.1
zipp==3.1.0
zmq==0.0.0
